Performance overhaul using Fiona.
Fiona uses a spatial indexing system to automatically filter objects based on a bounding box. This is really really faster than the geopandas system. Moreover, the lazy loading using Fiona reduces memory consumption as we do not have to load the full shapefile in memory, yay!

We switch the loading of the shapefiles to Fiona. It delegates to Geopandas only for the final stage, when computing intersecting areas of polygons with the raster.

Finally, we use GTiff compression to reduce the final raster size.